pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
]
